layout/generic/test/test_bug671319.html

<!DOCTYPE>
<html>
<head>
<title>test for bug 671319</title>
<script type="text/javascript"
  src="chrome://mochikit/content/MochiKit/packed.js"></script>
<script type="text/javascript"
  src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<script type="text/javascript"
  src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
<link rel="stylesheet" type="text/css"
  href="chrome://mochikit/content/tests/SimpleTest/test.css" />
</head>
<body>
<iframe id="iframe1"
  src="data:text/html,&lt;input id='childInput' style='width: 100px; height: 20px;'&gt;"
  style="margin: 10px; width: 150px; height: 40px;"></iframe><br>
<iframe id="iframe2"
  src="data:text/html,&lt;input id='childInput' style='width: 100px; height: 20px;'&gt;"
  style="margin: 10px; width: 150px; height: 40px;"></iframe>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();

function doTest() {
  var iframe1 = document.getElementById("iframe1");
  var childInput1 = iframe1.contentDocument.getElementById("childInput");
  var childWindow1 = iframe1.contentWindow;

  var iframe2 = document.getElementById("iframe2");
  var childInput2 = iframe2.contentDocument.getElementById("childInput");
  childInput2.onmousedown = function (aEvent) {
    aEvent.preventDefault();
  };
  childInput2.onmousemove = function (aEvent) {
    childInput2.setCapture();
  };
  var childWindow2 = iframe2.contentWindow;

  synthesizeMouse(childInput1, 10, 10, { type: "mousedown", button: 0 }, childWindow1);
  iframe1.contentDocument.releaseCapture();
  synthesizeMouse(childInput2, 10, 10, { type: "mousedown", button: 0 }, childWindow2);
  synthesizeMouse(childInput2, 20, 10, { type: "mousemove" }, childWindow2);

  ok(true, "not crashed");

  SimpleTest.finish();
}

SimpleTest.waitForFocus(doTest, window);

</script>
</pre>
</body>
</html>
